Fix the case when inputting a new filename while finding a current resource
If the `projectile-completion-system` allows it when finding a current
resource user is able to input a non-existant filename. This filename
will be used to create a new buffer for given filename.
If the comletion system does not allow of inputting arbitrary filename
user will be placed in the dired buffer for given directory.

Closes #66