Define the projectile-rails-global-mode
Since
https://github.com/bbatsov/projectile/commit/6986e435e64e2f4ee43bf8093c11e1da5fb266fa
projectile is a global mode that means that `projectile-mode-hook` is
run only once when the mode is enabled. So we have to have a global mode
also. The global mode is calling `projectile-rails-on` for each buffer
to see if the local `projectile-rails-mode` should be turned on (if it's
a Rails project).

The new init.el:

(add-hook 'projectile-mode-hook 'projectile-rails-global-mode)

or simply:

(projectile-rails-global-mode)

The latter will also work for the old projectile.

Closes #105