Require cl
This fixes a problem in my setup where I get errors such as this one
when starting projectile-rails:

```
Error in post-command-hook (projectile-rails-global-mode-check-buffers): (void-variable for)
```

I honestly don't know enough about elisp to know if this is a proper fix
for everyone. But to me it looks like the use of the `for` and `loop`
keywords are failing because `cl` is not explicitly required.

Looking at the current Travis breakage, I would expect this change to
fix that too.