Follow the changes to projectile's `projectile-dir-files`
The function returns now the filenmaes relative to the passed directory.
I had to change function `projectile-rails-choices` so it returns a hast table
with values that have the directory path concatenated to the filenames.
Also in cases where the function needs to look at different directories a third
element to the DIRS argument can be passed. If present the choices in minibuffer
will be "prefix + filename". So for `projectile-rails-find-environment` the
choices will be:
* environment
* application
* config/development

Instead of:
* environment
* application
* development

Not all tests are passing. But I prefer to have an unstable version compatible
with the upstream projectile than not working at all. The projectile internals
are changing dynamically right now so maybe it's not worth to make the tests
green just yet.

Related to #128