Only expand rails snippets in rails projects
While most of these do not apply, I'm finding the `require
"rails_helper"` in the spec skeleton a bit jarring when I'm creating
spec files in non-Rails Ruby projects. Granted, this snippet allows me
to specify which helper gets `require`'d, but I think I'd rather have
a hook that gave me good defaults for non-Rails projects when I'm not
in one, and this feels like the wrong place to introduce nuances like
that.